Replaced default main with Rabeel's modified main.cpp file
